# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Rokwire Groups Building Block API Documentation.",
        "title": "Rokwire Groups Building Block API",
        "contact": {},
        "version": "1.29.0"
    },
    "host": "localhost",
    "basePath": "/gr",
    "paths": {
        "/admin/authman/synchronize": {
            "post": {
                "security": [
                    {
                        "AppUserAuth": []
                    }
                ],
                "description": "Synchronizes Authman groups membership",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "operationId": "AdminSynchronizeAuthman",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/admin/group/events/v3": {
            "post": {
                "security": [
                    {
                        "AppUserAuth": []
                    }
                ],
                "description": "Create a calendar event and link it to multiple group ids",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "operationId": "AdminCreateCalendarEventMultiGroup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.createCalendarEventMultiGroupData"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.createCalendarEventMultiGroupData"
                        }
                    }
                }
            }
        },
        "/api/admin/group/{group-id}/event/{event-id}": {
            "delete": {
                "security": [
                    {
                        "AppUserAuth": []
                    }
                ],
                "description": "Deletes a group event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "operationId": "AdminDeleteGroupEvent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "event-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/admin/group/{group-id}/events": {
            "get": {
                "security": [
                    {
                        "AppUserAuth": []
                    }
                ],
                "description": "Gives the group events.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "operationId": "AdminGetGroupEvents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/group/{group-id}/events/v3/load": {
            "post": {
                "security": [
                    {
                        "AppUserAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Gets the group calendar events",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "operationId": "AdminGetGroupCalendarEventsV3",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body data",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/GroupEventFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/admin/group/{group-id}/members/v2": {
            "post": {
                "security": [
                    {
                        "AppUserAuth": []
                    }
                ],
                "description": "Gets the list of group members.",
                "consumes": [
                    "text/plain"
                ],
                "tags": [
                    "Admin"
                ],
                "operationId": "AdminGetGroupMembersV2",
                "parameters": [
                    {
                        "description": "body data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MembershipFilter"
                        }
                    },
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/GroupMembership"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/group/{groupId}/posts": {
            "post": {
                "security": [
                    {
                        "AppUserAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "creates a post within the desired group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "operationId": "AdminCreateGroupPost",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Post"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "AccountIdentifiers": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string"
                }
            }
        },
...
        "rest.createCalendarEventMultiGroupData": {
            "type": "object",
            "properties": {
                "admins_identifiers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/AccountIdentifiers"
                    }
                },
                "event": {
                    "type": "object",
                    "additionalProperties": true
                },
                "group_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "type": "apiKey",
            "name": "ROKWIRE-API-KEY",
            "in": "header"
        },
        "AppUserAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header (add Bearer prefix to the Authorization value)"
        },
        "IntAPIKeyAuth": {
            "type": "apiKey",
            "name": "INTERNAL-API-KEY",
            "in": "header"
        }
    }
}